fix: drop synapse_id when writing SONATA
Parquet has no implicit indexing, thus we write a column `synapse_id`
whenever we convert binary touches into Parquet.  This column has to be
present in any Parquet files to guarantee a unique sorting, but may be
dropped when writing SONATA (which has an implicit index).

Drop `synapse_id` as to not confuse end-users with internally used
columns.

Fixes FUNCZ-290.