FUNCZ-359 Replace synapse-tool dependency.
Add indexing based on libsonata to replace the synapse-tool dependency that we are not going to open source.

At the same time, simplify the code a bit by removing conditional compilation of MPI: within Spack, we only build with MPI.  This also helps with the indexing. Remove unused / obsolete test files that duplicate the main executables or are copied from Apache.