various fixes for mac (#2)
- INT64 parquet is a long on ubuntu and a long long on mac. Moved to int64_t
- `join_with_view` just does not stop complaining about forward range even if vector has it. I suspect it does not support something and I do not have cpp 23. Let's go back to the world of streams for now. The error is related to a `forward_range` not defined. I tried also with vectors and so on and it keeps giving the same error. I could not find a way to fix this except reverting to the old ways


other similar methods do not even appear to exist in `range/v3`. I think we are using something that included in the standard in cpp 23 but we are using the old proposed version. It is hard to find documentation for ranges/v3 as it is shadowed by the std documentation.

Also, `join_with` looks like is missing. Sincerely, this version of ranges is a little faulty and spotty. Do we really want this dependency when it is not needed?


- various smaller problems, I forgot

fix #1

Tested using spack